refactor: Transaction.Amount を decimal から int に変更
日本円は小数点以下を持たないため、型レベルで不正値（0.5円など）を
防げる int に変更。numeric(12,0) より integer の方が意図が明確。

- Domain/Application/Endpoints 全層で decimal → int に統一
- EF Core 設定: HasColumnType("numeric(12,0)") → "integer"
- マイグレーション追加: ChangeTransactionAmountToInteger
- テスト: decimal リテラル(m サフィックス)を int リテラルに修正
- docs/db-design.md: ER図・テーブル定義・設計判断セクションを更新

Subscription.Amount は別ブランチスコープのため今回は対象外。

+++ backend/src/KakeiBase.WebApi/Application/DTOs/Transactions/TransactionDto.cs
@@ -6,7 +6,7 @@ public record TransactionDto(
     Guid Id,
     Guid CategoryId,
     Guid? SubscriptionId,
-    decimal Amount,
+    int Amount,
     TransactionType Type,
     DateOnly Date,
     string? Memo,
