+++ backend/src/KakeiBase.WebApi/Endpoints/Requests/UpdateTransactionRequest.cs
@@ -5,7 +5,7 @@ namespace KakeiBase.WebApi.Endpoints.Requests;
 
 public record UpdateTransactionRequest(
     Guid CategoryId,
-    decimal Amount,
+    int Amount,
     TransactionType Type,
     DateOnly Date,
     string? Memo,
